refactor: extract dxf layout objects
