+++ plugins/CMakeLists.txt
@@ -16,6 +16,7 @@ add_library(cadgf_dxf_importer_plugin SHARED
     dxf_parser_zero_record.cpp
     dxf_header_vars.cpp
     dxf_parser_name_routing.cpp
+    dxf_layout_objects.cpp
     dxf_math_utils.cpp
     dxf_text_encoding.cpp
     dxf_color.cpp
